pyproject.toml
[tool.poetry]
name = "diagnosis"
version = "0.1.0"
description = ""
authors = ["se7en <gideon.ahiadzi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
gevent = "^22.10.2"
drf-yasg = "^1.21.5"
django-cors-headers = "^4.0.0"
whitenoise = "^6.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
coverage = "^7.2.7"

[tool.black]
line-length = 89
exclude = "migrations/*"

[tool.isort]
profile = "black"
line_length = 89
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
group_by_package = true
skip_glob = ["*/migrations/*.py"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "diagnosis.settings.testing"
pythonpath = [".", "src"]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
